backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httptools==0.6.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
aiohttp==3.11.0
tenacity==9.0.0
sentence-transformers==3.3.1
transformers==4.47.1
huggingface-hub==0.27.0
tokenizers==0.21.0
google-generativeai==0.8.3
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
langdetect==1.0.9
deep-translator==1.11.4
faiss-cpu
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
cachetools==5.5.0
pydantic==2.10.3
redis==5.2.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
requests==2.32.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
slowapi==0.1.9
scikit-learn==1.5.2
scipy==1.15.0
